backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

python-dotenv==1.0.1

pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.27.2

python-jose[cryptography]==3.3.0
